Sellar el run completo en el checkpoint y cerrar la Fase 17
Tres correcciones de una revision externa, las tres ciertas, y la prueba
de reanudacion con el modelo real.

Reanudar no respetaba su contrato. El comentario decia que manda el
checkpoint, pero semilla y n_validacion se comparaban contra los valores
por defecto del programa, asi que reanudar sin repetirlos abortaba. Y el
problema era mas amplio: iteraciones, lote, lr o el calendario no se
comprobaban, y reanudar con otro total cambiaba el coseno sin avisar.
Ahora todo lo que decide la trayectoria va sellado en los tres archivos,
se lee con nsf::ReadMetadata antes de construir nada, se adopta si no se
pasa y aborta si se pasa distinto. Reanudar sin repetir ninguna opcion da
pesos y EMA identicos a no parar.

sample_diffusion reconstruia el calendario de memoria; con otro, los
pesos cargan igual y las muestras salen peores sin error. Ahora pasos,
beta_ini, beta_fin y la normalizacion van en el sello y se adoptan; pedir
otro aborta, y un checkpoint sin sello avisa.

DDIM aceptaba un solo paso, que genera la subsecuencia {0} y presenta
ruido puro al modelo como si fuera t=0. El Test 54 eximia ese caso en vez
de detectarlo. Ahora exige al menos dos pasos cuando T > 1 y la prueba
comprueba que protesta.

De paso, EncodePngGris usaba static_cast<long> para desplazar iteradores,
de 32 bits en Windows; pasa a std::ptrdiff_t.

La reanudacion con el modelo real, desde unet_mnist_it070000 hasta la
80 000 con los argumentos del run original: pesos vivos, sombra de la EMA
y momentos m y v de Adam identicos a release/unet_mnist.nsf, con los
mismos contadores y la misma evaluacion final. El binario habia cambiado
desde el run, la particion se leyo del checkpoint antiguo como contigua y
durante el run se compilaba en paralelo; nada movio un bit.

Con eso la Fase 17 no tiene casillas abiertas. Roadmap: 115 cerrados, 8
pendientes.

+++ src/image/png.cpp
@@ -6,6 +6,7 @@
 
 #include "image/png.h"
 #include <algorithm>
+#include <cstddef>
 #include "image/inflate.h"
 
 namespace neuralsuite {
@@ -331,8 +332,8 @@ bool EncodePngGris(const Bitmap& in, std::vector<uint8_t>* out, std::string* err
   for (int y = 0; y < in.height; ++y) {
     crudo.push_back(0);
     const size_t base = static_cast<size_t>(y) * static_cast<size_t>(in.width);
-    crudo.insert(crudo.end(), in.pixels.begin() + static_cast<long>(base),
-                 in.pixels.begin() + static_cast<long>(base + static_cast<size_t>(in.width)));
+    crudo.insert(crudo.end(), in.pixels.begin() + static_cast<std::ptrdiff_t>(base),
+                 in.pixels.begin() + static_cast<std::ptrdiff_t>(base + static_cast<size_t>(in.width)));
   }
 
   // Envoltura zlib con bloques DEFLATE almacenados, de 65 535 bytes como mucho:
@@ -347,8 +348,8 @@ bool EncodePngGris(const Bitmap& in, std::vector<uint8_t>* out, std::string* err
     zlib.push_back(static_cast<uint8_t>(len >> 8));
     zlib.push_back(static_cast<uint8_t>(~len & 0xFF));
     zlib.push_back(static_cast<uint8_t>((~len >> 8) & 0xFF));
-    zlib.insert(zlib.end(), crudo.begin() + static_cast<long>(pos),
-                crudo.begin() + static_cast<long>(pos + len));
+    zlib.insert(zlib.end(), crudo.begin() + static_cast<std::ptrdiff_t>(pos),
+                crudo.begin() + static_cast<std::ptrdiff_t>(pos + len));
     pos += len;
   } while (pos < crudo.size());
   AnadirBigEndian32(&zlib, detail::Adler32(crudo.data(), crudo.size()));
